+++ src/main.js
@@ -3,12 +3,18 @@
 import { Home } from './components/Home.js';
 import { Register } from './components/Register.js';
 import { Login } from './components/Login.js';
+import { Welcome } from './components/Welcome.js';
+import { Publications } from './components/Publications.js';
+import { Profile } from './components/Profile.js';
 
 const rootDiv = document.getElementById('root');
 const routes = {
-  '/': Home,
+  '/': Welcome,
   '/register': Register,
   '/login': Login,
+  '/publications': Publications,
+  '/profile': Profile,
+  '/Home': Home,
 };
 /* const component = routes[window.location.pathname];
 rootDiv.appenChild(component()); */
